python/agent/langgraph-otel/pyproject.toml
[project]
name = "langgraph-otel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.6.0",
    "langgraph-prebuilt>=0.6.0",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    "opentelemetry-exporter-otlp>=1.36.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.36.0",
    "opentelemetry-exporter-otlp-proto-http>=1.36.0",
    "python-dotenv>=1.1.0",
]
